Adding Synchronization
Added Synchronization code to the new Project. Hasn't been totally
enabled as of yet, however the core framework to make it work. A few
small code changes in database adaptor and sync adaptor.

Initial Synchronization has been added to the login activity, so that it
occurs at the same time as the loading screen (assuming it works...
which it should)