Sync Adapter Functional
Sync Adapter is currently functional.
Going to go test the update table really quick in the other program.
Sync is currently forced on creation of a new account (first time sign
on, and only the first), and then after that should occur according to
android device choice.

I also have it set up so that every 15 times it synchronizes,  it will
(after synchronizing) drop the tables and reload from the database. This
should ensure parity between the two distinct servers, without any
additional overhead, as sync tends to only happen when the device is
charging / on wi-fi.